+++ proxy/internal/hannah/version_interceptor.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"strconv"
 
-	pb "github.com/NurPech/hannah-proto-go"
+	pb "github.com/NurPech/hannah-proto-go/v2"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 )